Render and preview rich text
Allow users to use rich text *bold* _italic_ ~strikethrough~ and
combinations of them in sent messages and preview this in the message
input field.

Use regular expressions to detect rich text styling. Multiple regular
expressions instead of a 'universal' rich text expression because the
latter would lead to more backtracking and potentially to 'catastrophic
backtracking'.

Use <strong> for bold, <em> for italic, and <s> for strikethrough to
render rich text styling. <strong> instead of changing the CSS font
weight because the former indicates text of special importance, which is
what a user using asterisks is implying. <em> instead of changing the
CSS font style because the former indicates stressed emphasis, which is
what a user using underscores is implying. <s> because it indicates text
that is no longer relevant/accurate, which is what a user using tildes
is implying. Additional CSS to ensure screen readers notify users that
text is stricken as it would otherwise adversely affect understanding.

Use the 'text-change' Event of Quill to check for and apply rich text
changes as a user types into the message input field by re-analyzing the
entire message input. This could possibly be improved by somehow only
analyzing changes to the input field, but the complexity of such a
solution might not be worth it.

To add more styles in the future, it only has to be added to the
Stylify.STYLES variable, and any Quill formats it applies must be
enabled on the <ReactQuill /> component - the regular expressions etc.
are handled by StyleType automatically.

Note that this only implements single-character styles e.g. '*'. This is
done to reduce how much a user has to type, and ease transitions from
other mobile messaging apps. Multi-character styles would require a
refactor.